+++ mmdet/core/__init__.py
@@ -1,6 +1,6 @@
 from .anchor import *  # noqa: F401, F403
-from .bbox_ops import *  # noqa: F401, F403
-from .mask_ops import *  # noqa: F401, F403
+from .bbox import *  # noqa: F401, F403
+from .mask import *  # noqa: F401, F403
 from .losses import *  # noqa: F401, F403
 from .eval import *  # noqa: F401, F403
 from .parallel import *  # noqa: F401, F403
