+++ mmdet/core/bbox/geometry.py
